+++ _examples/mvc/versioned-controller/main_test.go
@@ -12,21 +12,21 @@ func TestVersionedController(t *testing.T) {
 	app := newApp()
 
 	e := httptest.New(t, app)
-	e.GET("/data").WithHeader(versioning.AcceptVersionHeaderKey, "1").Expect().
+	e.GET("/data").WithHeader(versioning.AcceptVersionHeaderKey, "1.0.0").Expect().
 		Status(iris.StatusOK).Body().Equal("data (v1.x)")
 	e.GET("/data").WithHeader(versioning.AcceptVersionHeaderKey, "2.3.0").Expect().
 		Status(iris.StatusOK).Body().Equal("data (v2.x)")
-	e.GET("/data").WithHeader(versioning.AcceptVersionHeaderKey, "3.1").Expect().
+	e.GET("/data").WithHeader(versioning.AcceptVersionHeaderKey, "3.1.0").Expect().
 		Status(iris.StatusOK).Body().Equal("data (v3.x)")
 
 	// Test invalid version or no version at all.
-	e.GET("/data").WithHeader(versioning.AcceptVersionHeaderKey, "4").Expect().
+	e.GET("/data").WithHeader(versioning.AcceptVersionHeaderKey, "4.0.0").Expect().
 		Status(iris.StatusOK).Body().Equal("data")
 	e.GET("/data").Expect().
 		Status(iris.StatusOK).Body().Equal("data")
 
 	// Test Deprecated (v1)
-	ex := e.GET("/data").WithHeader(versioning.AcceptVersionHeaderKey, "1.0").Expect()
+	ex := e.GET("/data").WithHeader(versioning.AcceptVersionHeaderKey, "1.0.0").Expect()
 	ex.Status(iris.StatusOK).Body().Equal("data (v1.x)")
 	ex.Header("X-API-Warn").Equal(opts.WarnMessage)
 	expectedDateStr := opts.DeprecationDate.Format(app.ConfigurationReadOnly().GetTimeFormat())