+++ _examples/sessions/securecookie/main_test.go
@@ -16,12 +16,12 @@ func TestSessionsEncodeDecode(t *testing.T) {
 	es.Cookies().NotEmpty()
 	es.Body().Equal("All ok session set to: iris [isNew=true]")
 
-	e.GET("/get").Expect().Status(iris.StatusOK).Body().Equal("The name on the /set was: iris")
+	e.GET("/get").Expect().Status(iris.StatusOK).Body().Equal("The username on the /set was: iris")
 	// delete and re-get
 	e.GET("/delete").Expect().Status(iris.StatusOK)
-	e.GET("/get").Expect().Status(iris.StatusOK).Body().Equal("The name on the /set was: ")
+	e.GET("/get").Expect().Status(iris.StatusOK).Body().Equal("The username on the /set was: ")
 	// set, clear and re-get
 	e.GET("/set").Expect().Body().Equal("All ok session set to: iris [isNew=false]")
 	e.GET("/clear").Expect().Status(iris.StatusOK)
-	e.GET("/get").Expect().Status(iris.StatusOK).Body().Equal("The name on the /set was: ")
+	e.GET("/get").Expect().Status(iris.StatusOK).Body().Equal("The username on the /set was: ")
 }